Kullanıcı Sorularım Ekranı Puan Verme Backend Tamamlandı  #12

+++ BitirmeProjesi.Data/Questions.cs
@@ -19,7 +19,7 @@ namespace BitirmeProjesi.Data
         public string Description { get; set; }
         public Nullable<long> UserId { get; set; }
         public Nullable<long> VetId { get; set; }
-        public Nullable<decimal> RatingScore { get; set; }
+        public Nullable<int> RatingScore { get; set; }
         public Nullable<System.DateTime> CreateDate { get; set; }
     
         public virtual Users Users { get; set; }
